+++ examples/helper.py
@@ -1,5 +1,0 @@
-import os
-
-ENDPOINT = os.environ['ENDPOINT']
-DIR_PATH = os.path.dirname(os.path.realpath(__file__))
-ABI_FILEPATH = os.path.join(DIR_PATH, os.pardir, 'abi.json')