+++ CHANGELOG.md
@@ -13,13 +13,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - This CHANGELOG file.
 - Multithreading tests.
+- Transactions Manager support.
 
 ### Changed
 
 - All `web3` related functions moved from `skale.utils.helper` to `skale.utils.web_utils`.
 - `skale.utils.helper.await_receipt` renamed to `skale.utils.web3_utils.wait_receipt`.
 - `get_provider` method moved from `Skale` class to the `web3_utils`.
-- `sign_and_send` method now works with `web3` object instead of `skale`.
+- `sign_and_send` signature is changed - wallet param is now optional. 
 - `@format` decorator renamed to `@format_fields`.
 - Updated tests.